(fix) формат даты

+++ src/app/model/repo-row.ts
@@ -1,3 +1,5 @@
+// Дату и время вывести в формате- день, месяц, год, часы, минуты.
+
 export class RepoRow {
     name: string;
     language: string;
@@ -5,7 +7,7 @@ export class RepoRow {
     isArch: string;
     linkUrl: string;
 
-    static dateOptions:any = {day: 'numeric', month: 'long', year: 'numeric', hour: 'numeric', minute: 'numeric'};
+    static dateOptions:any = {day: 'numeric', month: 'numeric', year: 'numeric', hour: 'numeric', minute: 'numeric'};
 
     constructor(name: string, language: string, dateOfLastPush: string, isArch: string, linkUrl: string) {
         this.name = name;
